requirements.txt
openai
python-dotenv
python-magic 
Pillow 
piexif 
exifread 
PyPDF2 
mutagen